last version of 0.1.3
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/ann@3 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/ANN/src/files/ANN.cpp
@@ -133,7 +133,7 @@ double ArtificialNeuralNetwork::getMeanSquareError(double* target){
 ////////////////////////////////////////////
 double ArtificialNeuralNetwork::sigmoid(double inValue){
 	
-	return (double)(1/(1+exp(-inValue)));
+	return 1.0/(1+exp(-inValue));
 
 }
 ////////////////////////////////////////////
